Add SSO region config and handle SearchMetadata
- AWS Identity manager can be configured in any region, so change the
region to be user configurable, defaulting to us-east-1 for backwards
compat. This is referred to as ```sso_region``` in ```~/.aws/config```.
May be worth following suit to prevent confusion?

- In my configuration the appinstances json response contains
(abbreviated) ```{"name": "$ACCT_ID ($ACCT_DESC)", "searchMetadata": {
"AccountId": "$ACCT_ID", "AccountName": "$ACCT_DESC",
"AccountEmail":"$ACCT_EMAIL" }}```, so the account lookup was changed to
try the ```name``` first, and then ```searchMetadata.AccountId``` if not
found.